+++ CHANGELOG.md
@@ -48,6 +48,7 @@
 - Reports now handle large date ranges more robustly
 - Unselecting a radio box by clicking on it now properly applies any conditional branching logic
 - Hitting the `Enter` key when creating or updating a domain on the design editor now correctly saves the domain
+- Changing a subject on a sheet that is part of a schedule now correctly clears the schedule and event
 
 ## 0.21.3 (October 28, 2013)